Added logging and refactoring
The program now logs the movies in a file so the program is now persistent. On first launch the program will get all movies and use that as the current list to compare against. Main.py has been changed to .pyw so there is no console when running, and the chromedriver.exe no longer brings up a terminal each time